fix(config): make internal/config tests build on Windows
internal/config tests could not compile on Windows: config_test.go
called syscall.Mkfifo, which does not exist there. Because the failure
was at compile time, the package's existing t.Skipf fallback could
never run, and every test in the package was unreachable on Windows,
including the ones make docs-check and make stats depend on.

The Windows CI lane did not catch this: it runs go build ./... (which
never compiles _test.go files) plus targeted go test runs on five
other packages, none of which is internal/config.

Split the FIFO case by build tag:

  - license_nonregular_unix_test.go keeps the FIFO case, which is the
    one resolveLicenseKey specifically guards against, since a FIFO can
    block the reader.
  - license_nonregular_windows_test.go uses a directory instead.
    Windows has no mkfifo, but a directory is equally non-regular, so
    it exercises the same Mode().IsRegular() rejection rather than
    dropping the coverage.
  - license_nonregular_test.go holds the shared assertion, since both
    platforms are asserting the same security property.

Add a scoped step to the Windows lane so this cannot regress silently.
Compiling the package is the actual guard; go build ./... cannot
provide it because it never builds _test.go files. The stale lane
comment claiming internal/config uses unix-only syscalls is updated.

Also force LF in the working tree for *.md and *.go. Git for Windows
defaults core.autocrlf=true, and a CRLF checkout breaks two gates with
no real content drift:

  - TestDocsDeclareLiveStatsAndDefaults matches multi-line literals
    containing \n (docs_accuracy_test.go:26), so it failed on every
    checked doc.
  - gofumpt treats a CRLF file as unformatted, so make lint reported
    every Go file in the repo as needing formatting.

This follows the existing **/testdata/** rule, which is in
.gitattributes for the same reason. No committed .go or .md blob
contains CRLF today, so these rules normalize nothing and produce no
renormalization diff; on Linux they are a no-op.

With these changes make docs-check passes on Windows.

Scope note: this makes the package compile and fixes the doc and lint
gates. 27 tests in internal/config still fail on Windows because they
assert Unix permission-model behavior (world-readable CA key,
group-writable license_file, owner-executable 0o700, private temp
base). Those need either explicit Windows skips or ACL equivalents,
which is a separate decision and is left untouched here.
